add no-std CI check
