+++ Cargo.toml
@@ -63,7 +63,7 @@ jmt = "0.8.0"
 async-trait = "0.1.71"
 anyhow = { version = "1.0.68", default-features = false }
 arbitrary = { version = "1.3.1", features = ["derive"] }
-borsh = { version = "0.10.3", default-features = false, features = ["rc", "bytes"] }
+borsh = { version = "0.10.3", default-features = false }
 # TODO: Consider replacing this serialization format
 #     https://github.com/Sovereign-Labs/sovereign-sdk/issues/283
 bincode = "1.3.3"
@@ -80,10 +80,9 @@ proptest-derive = "0.3.0"
 rand = "0.8"
 rayon = "1.8.0"
 rocksdb = { version = "0.21.0", features = ["lz4"] }
-serde = { version = "1.0.188", default-features = false, features = ["alloc", "derive", "rc"] }
+serde = { version = "1.0.188", default-features = false, features = ["alloc", "derive"] }
 serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
 sha2 = { version = "0.10.6", default-features = false }
-spin = "0.9.8"
 thiserror = "1.0.50"
 tiny-keccak = "2.0.2"
 tracing = { version = "0.1.40", default-features = false }
