+++ Makefile
@@ -1,3 +1,6 @@
+# no-std packages to be checked
+NO_STD := "sov-rollup-interface"
+
 .PHONY: help
 
 help: ## Display this help message
@@ -23,6 +26,7 @@ install-dev-tools:  ## Installs all necessary cargo helpers
 	cargo install cargo-nextest --locked
 	cargo install cargo-risczero
 	cargo risczero install
+	rustup target add thumbv6m-none-eabi
 
 lint:  ## cargo check and clippy. Skip clippy on guest code since it's not supported by risc0
 	## fmt first, because it's the cheapest
@@ -41,6 +45,18 @@ check-features: ## Checks that project compiles with all combinations of feature
 
 check-fuzz: ## Checks that fuzz member compiles
 	$(MAKE) -C fuzz check
+
+check-no-std: ## Checks that project compiles without std
+	@for package in $(NO_STD); do \
+		echo "Checking no-std $${package}..."; \
+		cargo check -p $$package \
+			--target thumbv6m-none-eabi \
+			--no-default-features ; \
+		cargo check -p $$package \
+			--target thumbv6m-none-eabi \
+			--no-default-features \
+			--features native ; \
+	done
 
 find-unused-deps: ## Prints unused dependencies for project. Note: requires nightly
 	cargo udeps --all-targets --all-features
