+++ rollup-interface/src/state_machine/mocks/da.rs
@@ -2,16 +2,10 @@ use core::fmt::Display;
 use core::str::FromStr;
 
 use borsh::{BorshDeserialize, BorshSerialize};
-use bytes::Bytes;
 use serde::{Deserialize, Serialize};
 
-use crate::da::{
-    BlobReaderTrait, BlockHashTrait, BlockHeaderTrait, CountedBufReader, DaSpec, DaVerifier, Time,
-};
+use crate::da::{BlockHashTrait, BlockHeaderTrait, Time};
 use crate::maybestd::string::String;
-use crate::maybestd::vec::Vec;
-use crate::mocks::MockValidityCond;
-use crate::services::da::SlotData;
 use crate::{BasicAddress, RollupAddress};
 
 const JAN_1_2023: i64 = 1672531200;
@@ -112,60 +106,6 @@ impl Display for MockAddress {
 impl BasicAddress for MockAddress {}
 impl RollupAddress for MockAddress {}
 
-#[derive(
-    Debug,
-    Clone,
-    PartialEq,
-    borsh::BorshDeserialize,
-    borsh::BorshSerialize,
-    serde::Serialize,
-    serde::Deserialize,
-)]
-
-/// A mock BlobTransaction from a DA layer used for testing.
-pub struct MockBlob {
-    address: MockAddress,
-    hash: [u8; 32],
-    data: CountedBufReader<Bytes>,
-}
-
-impl BlobReaderTrait for MockBlob {
-    type Address = MockAddress;
-
-    fn sender(&self) -> Self::Address {
-        self.address
-    }
-
-    fn hash(&self) -> [u8; 32] {
-        self.hash
-    }
-
-    fn verified_data(&self) -> &[u8] {
-        self.data.accumulator()
-    }
-
-    fn total_len(&self) -> usize {
-        self.data.total_len()
-    }
-
-    #[cfg(feature = "native")]
-    fn advance(&mut self, num_bytes: usize) -> &[u8] {
-        self.data.advance(num_bytes);
-        self.verified_data()
-    }
-}
-
-impl MockBlob {
-    /// Creates a new mock blob with the given data, claiming to have been published by the provided address.
-    pub fn new(data: Vec<u8>, address: MockAddress, hash: [u8; 32]) -> Self {
-        Self {
-            address,
-            data: CountedBufReader::new(bytes::Bytes::from(data)),
-            hash,
-        }
-    }
-}
-
 /// A mock hash digest.
 #[derive(
     Clone,
@@ -241,62 +181,9 @@ impl BlockHeaderTrait for MockBlockHeader {
     }
 }
 
-/// A mock block type used for testing.
-#[derive(Serialize, Deserialize, PartialEq, core::fmt::Debug, Clone)]
-pub struct MockBlock {
-    /// The header of this block.
-    pub header: MockBlockHeader,
-    /// Validity condition
-    pub validity_cond: MockValidityCond,
-    /// Blobs
-    pub blobs: Vec<MockBlob>,
-}
-
-impl Default for MockBlock {
-    fn default() -> Self {
-        Self {
-            header: MockBlockHeader {
-                prev_hash: [0; 32].into(),
-                hash: [1; 32].into(),
-                height: 0,
-            },
-            validity_cond: Default::default(),
-            blobs: Default::default(),
-        }
-    }
-}
-
-impl SlotData for MockBlock {
-    type BlockHeader = MockBlockHeader;
-    type Cond = MockValidityCond;
-
-    fn hash(&self) -> [u8; 32] {
-        self.header.hash.0
-    }
-
-    fn header(&self) -> &Self::BlockHeader {
-        &self.header
-    }
-
-    fn validity_condition(&self) -> MockValidityCond {
-        self.validity_cond
-    }
-}
-
 /// A [`DaSpec`] suitable for testing.
 #[derive(serde::Serialize, serde::Deserialize, Debug, PartialEq, Eq)]
 pub struct MockDaSpec;
-
-impl DaSpec for MockDaSpec {
-    type SlotHash = MockHash;
-    type BlockHeader = MockBlockHeader;
-    type BlobTransaction = MockBlob;
-    type Address = MockAddress;
-    type ValidityCondition = MockValidityCond;
-    type InclusionMultiProof = [u8; 32];
-    type CompletenessProof = ();
-    type ChainParams = ();
-}
 
 /// The configuration for mock da
 #[derive(Debug, Clone, PartialEq, serde::Deserialize, serde::Serialize)]
@@ -308,26 +195,6 @@ pub struct MockDaConfig {
 #[derive(Clone, Default)]
 /// DaVerifier used in tests.
 pub struct MockDaVerifier {}
-
-impl DaVerifier for MockDaVerifier {
-    type Spec = MockDaSpec;
-
-    type Error = anyhow::Error;
-
-    fn new(_params: <Self::Spec as DaSpec>::ChainParams) -> Self {
-        Self {}
-    }
-
-    fn verify_relevant_tx_list(
-        &self,
-        _block_header: &<Self::Spec as DaSpec>::BlockHeader,
-        _txs: &[<Self::Spec as DaSpec>::BlobTransaction],
-        _inclusion_proof: <Self::Spec as DaSpec>::InclusionMultiProof,
-        _completeness_proof: <Self::Spec as DaSpec>::CompletenessProof,
-    ) -> Result<<Self::Spec as DaSpec>::ValidityCondition, Self::Error> {
-        Ok(Default::default())
-    }
-}
 
 #[cfg(test)]
 mod tests {
