+++ rollup-interface/src/state_machine/mocks/mod.rs
@@ -4,13 +4,17 @@
 mod da;
 #[cfg(all(feature = "native", feature = "tokio"))]
 mod service;
+#[cfg(feature = "std")]
+mod use_std;
 mod validity_condition;
 mod zk_vm;
 pub use da::{
-    MockAddress, MockBlob, MockBlock, MockBlockHeader, MockDaConfig, MockDaSpec, MockDaVerifier,
-    MockHash, MOCK_SEQUENCER_DA_ADDRESS,
+    MockAddress, MockBlockHeader, MockDaConfig, MockDaSpec, MockDaVerifier, MockHash,
+    MOCK_SEQUENCER_DA_ADDRESS,
 };
 #[cfg(all(feature = "native", feature = "tokio"))]
 pub use service::MockDaService;
+#[cfg(feature = "std")]
+pub use use_std::{MockBlob, MockBlock};
 pub use validity_condition::{MockValidityCond, MockValidityCondChecker};
 pub use zk_vm::{MockCodeCommitment, MockProof, MockZkvm};
