+++ rollup-interface/src/state_machine/mod.rs
@@ -5,6 +5,7 @@ pub mod da;
 pub mod stf;
 pub mod zk;
 
+#[cfg(feature = "std")]
 pub use bytes::{Buf, BufMut, Bytes, BytesMut};
 use serde::de::DeserializeOwned;
 use serde::Serialize;
